parallel done

+++ indexed_dataset.py
@@ -29,7 +29,7 @@ dtypes = {
     3: np.int16,
     4: np.int32,
     5: np.int64,
-    6: np.float,
+    6: float,
     7: np.double,
 }
